% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\docType{methods}
\name{CheckModelValidity}
\alias{CheckModelValidity}
\alias{CheckModelValidity,ConformationalSelectionModel-method}
\alias{CheckModelValidity,InducedFitModel-method}
\alias{CheckModelValidity,LangmuirModel-method}
\alias{CheckModelValidity,TwoStateModel-method}
\title{S4 generic help function to check model validity}
\usage{
CheckModelValidity(object)

\S4method{CheckModelValidity}{LangmuirModel}(object)

\S4method{CheckModelValidity}{InducedFitModel}(object)

\S4method{CheckModelValidity}{ConformationalSelectionModel}(object)

\S4method{CheckModelValidity}{TwoStateModel}(object)
}
\arguments{
\item{x}{the model object to be checked for}
}
\description{
generic function to be called by constructor for 
	model validity
}
\details{
it actually is called by different model method 
	setValidty(objects) to check for the validity
}
\section{Methods (by class)}{
\itemize{
\item \code{LangmuirModel}: check the validity of LangmuirModel

\item \code{InducedFitModel}: check the validity of InducedFitModel

\item \code{ConformationalSelectionModel}: check the validity of ConformationalSelectionMdoel

\item \code{TwoStateModel}: check the validity of TwoStateModel
}}

