% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\docType{class}
\name{ConformationalSelectionModel-class}
\alias{ConformationalSelectionModel-class}
\title{S4 class of the SPR ConformationalSelection model}
\description{
ConformationalSelection model object
}
\details{
check the manual for the detailed model specifications
	\url{http://}
}
\section{Slots}{

\describe{
\item{\code{BaseModel}}{a LangmuirModel object to hold the parameter
see \code{\link{LangmuirModel-class}}}

\item{\code{kf}}{numeric association rate constant}

\item{\code{kr}}{numeric dissociation rate constant}

\item{\code{R02}}{numeric vector the starting RUs for the dissocaiton phase.
optional}
}}
\seealso{
\code{\link{LangmuirModel-class}} 
	\code{\link{InducedFitModel-class}}
	\code{\link{TwoStateModel-class}}
}

