% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\name{ConformationalSelectionModel}
\alias{ConformationalSelectionModel}
\title{Constructor for \code{ConformationalSelectionModel}}
\usage{
ConformationalSelectionModel(kon, koff, kf, kr, analyteConcentrations, Rmax,
  associationLength, dissociationLength, R0 = numeric(), offset = 0)
}
\seealso{
\code{\link{ConformatinalSelectionModel-class}}
}

