% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\name{FitTwoStateSPR}
\alias{FitTwoStateSPR}
\title{fit the sensorgrame data based on the two state model}
\usage{
FitTwoStateSPR(x, mode = 1, type = 1, steadyStateStart = -1,
  steadyStateEnd = -1, windowSize = 10, init.association = NULL,
  init.dissociation = NULL, control = list(maxiter = 500, tol = 0.01,
  minFactor = 1/1e+10), trace = F, fix.ligand = TRUE, Rligand = NULL)
}
\arguments{
\item{x}{SensorgramData containing the data to be fitted}

\item{mode}{Integer to selection which type of fitting to do
1 to do nonlinear regression\cr
2 to do linear regression.\cr
nonlinear regression is better}

\item{type}{integer to indicate which type of analysis in dissociation
    phase to be used. 
1. for multi-state approximation
2. for two state do fitting and then regression.}

\item{steadyStateStart}{numeric the starting time for steady state
optional and if provided, will overwrite the one in sensorgramdata}

\item{steadyStateEnd}{numeric the ending time for steady state
optional and if provided, will overwrite the ones in sensorgramdata}

\item{windowSize}{numeric the time period used to do approximation to
estimate the mean dissociaiton rate constant}

\item{init.association}{list of initial values of parameter to do the 
non-linear regression. It has the following members
list(Rmax=200, KD=1E-3)}

\item{init.dissociation}{list of initial values of parameter to do the 
non-linear regression. It has the following members for type 1 analysis
list(R1=100, r1=0.1) and for type 2 analysis 
list(R1=20, R2=200,r1=-0.005,r2=-0.001)}

\item{control}{list of control elements for run non-linear regression}

\item{trace}{boolean to control whether show the trace of
nonlinear regression.}

\item{fix.ligand}{a boolean indicating which ligand immoblization
model is using. \cr
TRUE, the fixed ligand immobilzation model. Rmax is used\cr
    FALSE, the variable ligand immbolization model. Rligand 
and efficiency are used. See also \code{\link{LangmuirModel-class}}}

\item{Rligand}{the input for the variable immobilization levels of 
ligands on the chip. This is used by the variable ligand 
immbolization model.See also \code{\link{LangmuirModel-class}}}
}
\value{
a list of parameters estimated
}
\description{
estimate the mean equilibrium dissociation constant based on the two state model
}
\details{
It can also estimate the mean dissociation rate constants
	 at the steady state. It assumes the two state model, 
	but can be easily generilized to
	more than "two-conformation" cases. In the multi-conformation cases,
	the mean dissociation rate constant can only be approximated
	empirically using the data of a small window at the beginning of
	the dissociation phase. Check detail of the implementation here
	\url{http://}
	Note: this one is a S4 method. it is probably not necessary. We will
		change it(??).
}

