% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\name{GetObservedRUs}
\alias{GetObservedRUs}
\title{S3 function to combine/add two SensorgramData}
\usage{
GetObservedRUs(DATA1, DATA2)
}
\arguments{
\item{DATA1}{SensorgramData}

\item{DATA2}{SensorgramData}
}
\value{
a SensorgramData object with the sum of the two input objects
	on both association and dissociation data. It takes other slot
	values from DATA1 assuming DATA2 has the identical values on
	these other slots.
}
\description{
Combine two SensorgramData object to simulate
	the real readout on a SPR machine.
}
\details{
The idea is to combine the RUs from two different source
	e.g. from two complexes with different comfirmations, as
	an real readout on the SPR machine. It adds up the RUs at
	the same time points to get the sum. Do this through the 
	association and dissociation. Don't try to modify any other 
	slots for the object.
}
\seealso{
\code{\link{ReadSensorgramData}} \code{\link{SensorgramData}} \code{\link{plot}} \code{\link{SaveSPRData}}
}

