% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\docType{class}
\name{InducedFitModel-class}
\alias{InducedFitModel-class}
\title{S4 class of the SPR InducedFit model}
\description{
InducedFit model object
}
\details{
check the manual for the detailed model specifications
	\url{http://}
}
\section{Slots}{

\describe{
\item{\code{BaseModel}}{a LangmuirModel object to hold the parameter
see \code{\link{LangmuirModel}}}

\item{\code{kf}}{numeric association rate constant}

\item{\code{kr}}{numeric dissociation rate constant}

\item{\code{R02}}{numeric vector the starting RUs for the dissocaiton phase.
optional}
}}
\seealso{
\code{\link{LangmuirModel}} 
	\code{\link{ConformationalSelectionModel-class}}
	\code{\link{TwoStateModel-class}}
}

