% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\name{InducedFitModel}
\alias{InducedFitModel}
\title{Constructor for \code{InducedFitModel}}
\usage{
InducedFitModel(kon, koff, kf, kr, sd = 0, analyteConcentrations, Rmax,
  associationLength, dissociationLength, R0 = numeric(), R02 = numeric(),
  offset = 0, Rligand = "numeric", efficiency = "numeric")
}
\seealso{
\code{\link{InducedFitModel-class}}
}

