% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\docType{class}
\name{LangmuirModel-class}
\alias{LangmuirModel-class}
\title{S4 class of the SPR Langmuir model}
\description{
Langmuir model object
}
\details{
check the manual for the detailed model specifications
	\url{http://}
}
\section{Slots}{

\describe{
\item{\code{kon}}{numeric association rate constant}

\item{\code{koff}}{numeric dissociation rate constant}

\item{\code{analyteConcentrations}}{numeric vector for different
analyte concentrations.}

\item{\code{sd}}{numeric standard deviation of the Gaussian noise. 
by default sd=0, which there is no noise.}

\item{\code{Rmax}}{numeric the maximum RUs levels at the infinite
high analyte concentration.}

\item{\code{associationLength}}{numeric the time period to run the association}

\item{\code{dissociationLength}}{numeric the time period to run the dissociation}

\item{\code{R0}}{numeric vector the starting RUs for the dissocaiton phase.
optional}

\item{\code{offset}}{the level shifted at the begining of the dissociation phase
optional}
}}
\seealso{
\code{\link{ConformationalSelectionModel-class}}
	\code{\link{TwoStateModel-class}}
	\code{\link{InducedFitModel-class}}
}

