% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\name{LangmuirModel}
\alias{LangmuirModel}
\title{Constructor for LangmuirModel}
\usage{
LangmuirModel(kon, koff, sd = 0, analyteConcentrations, Rmax,
  associationLength, dissociationLength, R0 = numeric(), offset = 0,
  Rligand = numeric, efficiency = 1)
}
\seealso{
\code{\link{LangmuirModel}}
}

