% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\docType{methods}
\name{plot,SensorgramData-method}
\alias{plot,SensorgramData-method}
\title{Generic plot function for SensorgramData}
\usage{
\S4method{plot}{SensorgramData}(x, col = 1, lty = 1, lwd = 1, main = "",
  xlab = "time (min)", ylab = "RUs")
}
\arguments{
\item{SensorgramData}{the SensorgramData object to be plotted}
}
\description{
\code{plot} plot the sensorgram
}
\details{
Generic function to plot the sensorgram of SPR.
	Both the dissociation and association phase as well
	the steady state window.
}
\seealso{
\code{\link{ReadSensorgramData}} \code{\link{SensorgramData}} \code{\link{SaveSPRData}}
}

