% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\name{ReadSensorgramData}
\alias{ReadSensorgramData}
\title{S3 method to read in the data}
\usage{
ReadSensorgramData(file, associationPhaseStart = 0,
  associationPhaseEnd = -1, dissociationPhaseEnd = -1,
  steadyStateStart = -1, steadyStateEnd = -1,
  analyteConcentrations = numeric(0), skip = 0, header = TRUE,
  sep = "\\t")
}
\arguments{
\item{file}{the directory path to the input file with the
correct format as mentioned above}

\item{associationPhaseStart}{numeric indiciation the starting 
time for associationPhase}

\item{associationPhaseEnd}{numeric indiciation the starting 
time for associationPhase. This is actually the starting
time for dissociationPhase}

\item{dissociationPhaseEnd}{numeric indiciation the starting 
time for dissociationPhase}

\item{steadyStateStart}{numeric indicating the starting time of 
steady state (optional).}

\item{steadyStateEnd}{numeric indicating the ending time of 
steady state (optional).}

\item{analyteConcentrations}{numeric vector for analyte concentrations}

\item{skip}{numeric the options for control the text reading. indicating
number of line to skip}

\item{header}{boolean header for the table reading (T/F)}

\item{sep}{character for the delimiter of the text}
}
\value{
a sensorgram data object
}
\description{
It read in data as a txt file according to
   exported SPR file from SensiQ
}
\details{
Data file has the following format\cr
\tabular{llllllll}{
  Curve\tab	1\tab	2\tab		3\tab		4\cr	
  Analyte\tab	CBS\tab	CBS\tab		CBS\tab		CBS\cr	
 Ligand\tab	Ch 1\tab	Ch 1\tab		Ch 1\tab		Ch 1\cr	
 Conc\tab	0.00E+00\tab		5.00E-05\tab		0.00E+00\tab		0.00E+00\cr
Time1\tab	Data1\tab	Time2\tab	Data2\tab	Time3\tab	Data3\tab	Time4\tab	Data4\cr
-2.6\tab	0\tab	-2.6\tab	0.2839\tab	-2.6\tab	0\tab	-2.6\tab	0\cr
-2.4\tab	0\tab	-2.4\tab	0.3632\tab	-2.4\tab	0\tab	-2.4\tab	0\cr
-2.2\tab	0\tab	-2.2\tab	0.6978\tab	-2.2\tab	0\tab	-2.2\tab	0\cr
}
Each file contains the header about the information of the data
Data colums for the raw sensorgram data with header for each column
each "curve" has two column, one time and the other RUs. 
The number of curves are variable. Please check the example
data file in the example/ folder.
Note: after reading the dissociation phase is rescaled to 
start at time zero.!!! relative to the 
dissociationPhaseStart/associationPhaseEnd
}
\seealso{
\code{\link{SensorgramData}} \code{\link{plot}} \code{\link{SaveSPRData}}
}

