% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\docType{class}
\name{SensorgramData-class}
\alias{SensorgramData-class}
\title{S4 class definition of SensorgramData}
\arguments{
\item{associationData}{Dataframe holding the association phase data. 
It has a fixed format/template.
See \code{\link{ReadSensorgramData}} for details.}

\item{dissociationData}{Dataframe holding the dissociation phase data. 
It has a fixed format/template.
See \code{\link{ReadSensorgramData}} for details.}

\item{analyteConcentrations, }{numeric vector for analyteConcentrations 
of unit molar concentration (M/L)}

\item{steadyStateStart}{numeric indicating the starting time of 
steady state (optional).}

\item{steadyStateEnd}{numeric indicating the ending time of 
steady state (optional).}

\item{offset}{numeric indicating the RU levels shift at the
begining of dissociation phase. Not used for now.
will implement in the future release. (optional)}
}
\description{
\code{SensorgramData} define the S4 class of SensorgramData
}
\details{
defining the S4 class of the SensorgrameData.
This is the data structure to hold the sensor grame Data. 
It contains different slots for holding both association and 
   dissociation phase data.It also defines 
   the steady state window if there is one. Analyte concentrations 
   also included. Sometimes the data could 
   only have either association or dissociation phase alone.
}
\section{Slots}{

\describe{
\item{\code{associationData}}{data.frame}

\item{\code{dissociationData}}{data.frame}

\item{\code{analyteConcentrations}}{numeric vector}

\item{\code{steadyStateStart}}{numeric}

\item{\code{steadyStateEnd}}{numeric}

\item{\code{offset}}{numeric}
}}
\examples{
	dt<-new("SensorgramData",
	dissociationData=data.frame(time=1:5,RU=1:5, time2=1:5, RU=1:5))
	dt@associationData

	dataFile<-list.files(system.file("extdata", package="SPRATS"),
	 pattern = "CBSInhibitor_raw.txt", full.names=TRUE) 
	rawData<-ReadSensorgramData(dataFile, skip=4, 
	header=T, sep="\\t", associationPhaseEnd=80, dissociationPhaseEnd=150)

	plot(rawData)
}
\seealso{
\code{\link{ReadSensorgramData}} \code{\link{GetObservedRUs}} \code{\link{SaveSPRData}}
}
\concept{
SPR
}

