% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\docType{methods}
\name{Simulate}
\alias{Simulate}
\alias{Simulate,ConformationalSelectionModel-method}
\alias{Simulate,InducedFitModel-method}
\alias{Simulate,LangmuirModel-method}
\alias{Simulate,TwoStateModel-method}
\title{S4 Geric method to Simluate SPR data}
\usage{
Simulate(x, sampleFreq = 0.01, ...)

\S4method{Simulate}{LangmuirModel}(x, sampleFreq = 0.01)

\S4method{Simulate}{InducedFitModel}(x, sampleFreq = 0.01)

\S4method{Simulate}{ConformationalSelectionModel}(x, sampleFreq = 0.01)

\S4method{Simulate}{TwoStateModel}(x, sampleFreq = 0.01, timeStep = 0.01)
}
\arguments{
\item{sampleFreq}{numeric the time frequency to collect the SPR data}
}
\value{
a list of \code{\link{SensorgramData-class}} data object 
	holding the SPR data for different component in the system
}
\description{
Geric method to run the simulatoin based on the model
}
\details{
the each model has different biochemical dynamics to generate
	the data. The Langmuir, induced fit and conformational selection
	models simulate data analytically, but the two state model simulate
	through numerical integration. Please check the detail here
	\url{http://}

the output is the sensorgram data for AB, A and A_star
  warning though, the association data for A and A_star 
  are good. The dissociation data for A is the sum of A and A_star
  since we can simply tell them apart with only R observed (R is the AB levels)
  So we simply add them together and save them in A@dissociationData and 
  save nothing in A_star (no dissociation data).
Note: this part is a bit confusing. Need more work in the future.
}
\section{Methods (by class)}{
\itemize{
\item \code{LangmuirModel}: to simulate the SPR data based on a Langmuir Model

\item \code{InducedFitModel}: to simulate the SPR data based on a InducedFit Model

\item \code{ConformationalSelectionModel}: to simulate the SPR data based on a ConformationalSelection Model

\item \code{TwoStateModel}: to simulate the SPR data based on a TwoState Model
}}
\seealso{
\code{\link{ConformatinalSelectionModel-class}}
		\code{\link{TwoStateModel-class}}
		\code{\link{InducedFitModel-class}}
		\code{\link{LangmuirModel-class}}
		\code{\link{SensorgramData-class}}
}

