% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRATS.R
\docType{package}
\name{SPRATS-package}
\alias{SPRATS}
\alias{SPRATS-package}
\title{SPR Analyzer for the Two State model}
\description{
An R package to characterize the Antibody Antigen 
	interactions under the Two State model by 
	surface plasmon resonance
}
\details{
This package is defined to run analysis of SPR data 
	under the two state model to estimate mean equilibrium
	dissociation constant as well as the mean dissociation rate
	constant. It takes in the exported text input data. Output 
	is estimated by non-linear regression. The modelling can be
	generilized to multi-state conditions. 
	This package can also provide functions to run simulations
	for different models, such as Langmuir model, induced fit
	model, conformation selection model and two state model. 

		Please refer to the vignettes to see details.
}

