% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\docType{class}
\name{TwoStateModel-class}
\alias{TwoStateModel-class}
\title{S4 class of the SPR Two State model}
\description{
Two State model object
}
\details{
check the manual for the detailed model specifications
	\url{http://}
	Note, 
	in this two state model case, we always using the parameters for
	length of period\cr
	 Rmax\cr
	 concentration\cr
	 R0\cr
		offset\cr
	 from the inducedFitModel@LangmuirModel
}
\section{Slots}{

\describe{
\item{\code{BaseModel1}}{an InducedFitModel object to hold the parameter
see \code{\link{InducedFitModel-class}}}

\item{\code{BaseModel2}}{a ConformationalSelectionModel object to hold the parameter
see \code{\link{ConformationalSelectionModel-class}}}
}}
\seealso{
\code{\link{LangmuirModel-class}} 
	\code{\link{InducedFitModel-class}}
	\code{\link{ConformationalSelectionModel-class}}
}

