% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\name{TwoStateModel}
\alias{TwoStateModel}
\title{Constructor for \code{TwoStateModel}}
\usage{
TwoStateModel(kon, koff, kf, sd = 0, kr, kon2, koff2, kf2, kr2,
  analyteConcentrations, Rmax, associationLength, dissociationLength,
  R0 = numeric(), R02 = numeric(), offset = 0, Rligand = "numeric",
  efficiency = "numeric")
}
\seealso{
\code{\link{TwoStateModel-class}}
}

