% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRTwoState.R
\docType{class}
\name{LangmuirModel-class}
\alias{LangmuirModel-class}
\title{S4 class of the SPR Langmuir model}
\description{
Langmuir model object
}
\details{
check the manual for the detailed model specifications
	\url{http://}
}
\section{Slots}{

\describe{
\item{\code{kon}}{numeric association rate constant}

\item{\code{koff}}{numeric dissociation rate constant}

\item{\code{analyteConcentrations}}{numeric vector for different
analyte concentrations.}

\item{\code{s_d}}{numeric standard deviation of the Gaussian noise. 
by default s_d=0, which there is no noise.}

\item{\code{Rmax}}{numeric the maximum RUs levels at the infinite
high analyte concentration. It is also the maximum level
of ligand immobilized on the chip surface \cr}

\item{\code{associationLength}}{numeric the time period to run the association}

\item{\code{dissociationLength}}{numeric the time period to run the dissociation}

\item{\code{R0}}{numeric vector the starting RUs for the dissocaiton phase.
optional}

\item{\code{offset}}{the level shifted at the begining of the dissociation phase
optional}

\item{\code{Rligand}}{the levels of ligand immobilized on the chip surface. 
optional\cr
If this one is set, then it assumes a variable "Rmax" model. Where 
Rmax for different run is not fixed but rather different due to 
some factors
such as regeration in (Sensiq) or stochasticity (in BioRad Proteon). 
In this model, the immoblized ligand is not fixed from run to run,
but the effienciency or conversion constant is fixed.\cr
This is a vector contains same number of element as the analyteConc.}

\item{\code{efficiency}}{the conversion constant for a "variable" Rmax model 
optional\cr
see above for explanation.}
}}
\seealso{
\code{\link{ConformationalSelectionModel-class}}
	\code{\link{TwoStateModel-class}}
	\code{\link{InducedFitModel-class}}
}

